refactor: moved components in proper folder
